Fixes from review
