+++ homeassistant/components/mastodon/coordinator.py
@@ -2,31 +2,16 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from datetime import timedelta
 from typing import Any
 
 from mastodon import Mastodon
 from mastodon.Mastodon import MastodonError
 
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import LOGGER
-
-
-@dataclass
-class MastodonData:
-    """Mastodon data type."""
-
-    client: Mastodon
-    instance: dict
-    account: dict
-    coordinator: MastodonCoordinator
-
-
-type MastodonConfigEntry = ConfigEntry[MastodonData]
 
 
 class MastodonCoordinator(DataUpdateCoordinator[dict[str, Any]]):
@@ -39,7 +24,7 @@ class MastodonCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         )
         self.client = client
 
-    async def _async_update_data(self) -> dict:
+    async def _async_update_data(self) -> dict[str, Any]:
         try:
             account: dict = await self.hass.async_add_executor_job(
                 self.client.account_verify_credentials
