+++ tests/components/mastodon/snapshots/test_sensor.ambr
@@ -99,7 +99,7 @@
     'state': '178',
   })
 # ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none-entry]
+# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_posts-entry]
   EntityRegistryEntrySnapshot({
     'aliases': set({
     }),
@@ -113,7 +113,7 @@
     'disabled_by': None,
     'domain': 'sensor',
     'entity_category': None,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none',
+    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_posts',
     'has_entity_name': True,
     'hidden_by': None,
     'icon': None,
@@ -125,174 +125,24 @@
     }),
     'original_device_class': None,
     'original_icon': None,
-    'original_name': None,
+    'original_name': 'Posts',
     'platform': 'mastodon',
     'previous_unique_id': None,
     'supported_features': 0,
-    'translation_key': 'followers',
-    'unique_id': 'client_id_followers',
-    'unit_of_measurement': 'followers',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none-state]
-  StateSnapshot({
-    'attributes': ReadOnlyDict({
-      'friendly_name': 'Mastodon @trwnh@mastodon.social None',
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-      'unit_of_measurement': 'followers',
-    }),
-    'context': <ANY>,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none',
-    'last_changed': <ANY>,
-    'last_reported': <ANY>,
-    'last_updated': <ANY>,
-    'state': '821',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none_2-entry]
-  EntityRegistryEntrySnapshot({
-    'aliases': set({
-    }),
-    'area_id': None,
-    'capabilities': dict({
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-    }),
-    'config_entry_id': <ANY>,
-    'device_class': None,
-    'device_id': <ANY>,
-    'disabled_by': None,
-    'domain': 'sensor',
-    'entity_category': None,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none_2',
-    'has_entity_name': True,
-    'hidden_by': None,
-    'icon': None,
-    'id': <ANY>,
-    'labels': set({
-    }),
-    'name': None,
-    'options': dict({
-    }),
-    'original_device_class': None,
-    'original_icon': None,
-    'original_name': None,
-    'platform': 'mastodon',
-    'previous_unique_id': None,
-    'supported_features': 0,
-    'translation_key': 'following',
-    'unique_id': 'client_id_following',
-    'unit_of_measurement': 'accounts',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none_2-state]
-  StateSnapshot({
-    'attributes': ReadOnlyDict({
-      'friendly_name': 'Mastodon @trwnh@mastodon.social None',
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-      'unit_of_measurement': 'accounts',
-    }),
-    'context': <ANY>,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none_2',
-    'last_changed': <ANY>,
-    'last_reported': <ANY>,
-    'last_updated': <ANY>,
-    'state': '178',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none_3-entry]
-  EntityRegistryEntrySnapshot({
-    'aliases': set({
-    }),
-    'area_id': None,
-    'capabilities': dict({
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-    }),
-    'config_entry_id': <ANY>,
-    'device_class': None,
-    'device_id': <ANY>,
-    'disabled_by': None,
-    'domain': 'sensor',
-    'entity_category': None,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none_3',
-    'has_entity_name': True,
-    'hidden_by': None,
-    'icon': None,
-    'id': <ANY>,
-    'labels': set({
-    }),
-    'name': None,
-    'options': dict({
-    }),
-    'original_device_class': None,
-    'original_icon': None,
-    'original_name': None,
-    'platform': 'mastodon',
-    'previous_unique_id': None,
-    'supported_features': 0,
-    'translation_key': 'statuses',
-    'unique_id': 'client_id_statuses',
-    'unit_of_measurement': 'statuses',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_none_3-state]
-  StateSnapshot({
-    'attributes': ReadOnlyDict({
-      'friendly_name': 'Mastodon @trwnh@mastodon.social None',
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-      'unit_of_measurement': 'statuses',
-    }),
-    'context': <ANY>,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_none_3',
-    'last_changed': <ANY>,
-    'last_reported': <ANY>,
-    'last_updated': <ANY>,
-    'state': '33120',
-  })
-# ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_statuses-entry]
-  EntityRegistryEntrySnapshot({
-    'aliases': set({
-    }),
-    'area_id': None,
-    'capabilities': dict({
-      'state_class': <SensorStateClass.TOTAL: 'total'>,
-    }),
-    'config_entry_id': <ANY>,
-    'device_class': None,
-    'device_id': <ANY>,
-    'disabled_by': None,
-    'domain': 'sensor',
-    'entity_category': None,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_statuses',
-    'has_entity_name': True,
-    'hidden_by': None,
-    'icon': None,
-    'id': <ANY>,
-    'labels': set({
-    }),
-    'name': None,
-    'options': dict({
-    }),
-    'original_device_class': None,
-    'original_icon': None,
-    'original_name': 'Statuses',
-    'platform': 'mastodon',
-    'previous_unique_id': None,
-    'supported_features': 0,
-    'translation_key': 'statuses',
-    'unique_id': 'client_id_statuses',
+    'translation_key': 'posts',
+    'unique_id': 'client_id_posts',
     'unit_of_measurement': 'posts',
   })
 # ---
-# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_statuses-state]
+# name: test_sensors[sensor.mastodon_trwnh_mastodon_social_posts-state]
   StateSnapshot({
     'attributes': ReadOnlyDict({
-      'friendly_name': 'Mastodon @trwnh@mastodon.social Statuses',
+      'friendly_name': 'Mastodon @trwnh@mastodon.social Posts',
       'state_class': <SensorStateClass.TOTAL: 'total'>,
       'unit_of_measurement': 'posts',
     }),
     'context': <ANY>,
-    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_statuses',
+    'entity_id': 'sensor.mastodon_trwnh_mastodon_social_posts',
     'last_changed': <ANY>,
     'last_reported': <ANY>,
     'last_updated': <ANY>,
